requirements.txt
numpy
opencv-python
pyyaml
scipy
